Refactor message encoding/decoding to use varint instead of uint16
- Updated SessionClientMessage, SessionServerMessage, SessionUpdateMessage, SubscribeMessage, SubscribeOkMessage, and SubscribeUpdateMessage to use varint for message length encoding and decoding.
- Modified associated test cases to reflect changes in encoding and error handling for varint.
- Adjusted frame encoding in Go to utilize varint for length representation.
- Ensured consistency across message handling in both TypeScript and Go implementations.

+++ moqt/frame.go
@@ -66,10 +66,10 @@ func (f *Frame) Cap() int {
 	return cap(f.body)
 }
 
-// encode writes the frame in MOQ format: u16 length followed by payload.
+// encode writes the frame in MOQ format: varint length followed by payload.
 // The length is encoded into the header buffer to minimize allocations.
 func (f *Frame) encode(w io.Writer) error {
-	l := uint16(len(f.body))
+	l := uint64(len(f.body))
 	header, _ := message.WriteMessageLength(f.header[:0], l)
 	start := 8 - len(header)
 	copy(f.buf[start:], header)
