+++ moqt/internal/message/session_update.go
@@ -20,7 +20,7 @@ func (sum SessionUpdateMessage) Encode(w io.Writer) error {
 	b := pool.Get(msgLen + VarintLen(uint64(msgLen)))
 	defer pool.Put(b)
 
-	b, _ = WriteMessageLength(b, uint16(msgLen))
+	b, _ = WriteMessageLength(b, uint64(msgLen))
 	b, _ = WriteVarint(b, sum.Bitrate)
 
 	_, err := w.Write(b)
