+++ moqt/internal/message/subscribe_ok.go
@@ -21,7 +21,7 @@ func (som SubscribeOkMessage) Encode(w io.Writer) error {
 	b := pool.Get(msgLen + VarintLen(uint64(msgLen)))
 	defer pool.Put(b)
 
-	b, _ = WriteMessageLength(b, uint16(msgLen))
+	b, _ = WriteMessageLength(b, uint64(msgLen))
 
 	_, err := w.Write(b)
